requirements.txt
Flask
mysql-connector-python
python-dotenv
flasgger
werkzeug